MAKERCHIP macro var, disable urandom in TT7 build
After debugging with @stevehoover, we discovered the macros need to be
adjusted and ensure that we ensure the urandom macro does not get pulled
in like before this change. Added docs for the var so you know to set it
back to 1 when using MakerChip for design and simulation testing.